Display wechat avatar with donator's name
