+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useLayoutEffect } from "react";
 import _ from "lodash";
-import  moment from "moment";
+import moment from "moment";
 
 // 引入样式
 import "./assets/styles/common.scss";
@@ -8,21 +8,23 @@ import "./assets/styles/app.scss";
 
 
 // 引入页面组件
-import Header from './components/Header'
-import Footer from './components/Footer'
-import Volunteer from './components/Volunteer'
-import Sponser from './components/Sponser'
-import HelpUs from './components/HelpUs'
+import Header from './components/Header';
+import Footer from './components/Footer';
+import Volunteer from './components/Volunteer';
+import Sponser from './components/Sponser';
+import HelpUs from './components/HelpUs';
+import Donators from './components/Donators';
+import Requesters from './components/Requesters';
 
 // 图片资源、金数据
-import {images, JINSHUJU} from "./utils/contants"
+import { images, JINSHUJU } from "./utils/contants"
 
 // 接口请求
 import * as request from './request'
 
 // utils
 import { buildChart } from "./utils/chartUtils";
-import { WXShare, getCountryName, getCityName } from "./utils";
+import { WXShare } from "./utils";
 
 // 轮播插件
 import Slider from "react-slick";
@@ -59,6 +61,17 @@ const App = () => {
   const [filteredData, setFilteredData] = useState([]);
   const [supportInfoList, setSupportInfoList] = useState([]);
 
+  // Props to be passed to Donators
+  const configDonatorsComponent = {
+    domesticData,
+    donatorData
+  };
+
+  // Props to be passed to Requesters
+  const configRequestersComponent = {
+    filteredData
+  };
+
   const toggleModal = (index) => {
     setModalIsOpen(!modalIsOpen);
     setSelectedIndex(index);
@@ -66,7 +79,7 @@ const App = () => {
 
   const updateVolunteerData=(data)=>{
     if(data&&data.length>0){
-      setVolunteerCount(data.length)
+      setVolunteerCount(data.length);
       setVolunteerData(data)
     }
   }
@@ -162,6 +175,7 @@ const App = () => {
       .then((response) => {
           let { data, next } = response;
           const newData = tmpData.concat(data).filter((d) => d["field_13"] === "已审核");
+          // console.log(newData);
           setDonatorData(newData.reverse());
           setMoney(
             newData.reduce((total, currentValue) => {
@@ -186,85 +200,6 @@ const App = () => {
     })
     .catch((error) => console.log(error));
   }, []);
-
-  // For donator table scrolling effect
-  useEffect(() => {
-    const donatorWrapper = document.querySelector("#donator-wrapper");
-    const donatorTable = document.querySelector(".info-table");
-
-    const requesterWrapper = document.querySelector("#requester-wrapper");
-
-    const donatorScroll = () => {
-      // When donator list is scrolled to the end, the cycle will restart
-      if (donatorWrapper.scrollTop >= 0.5 * donatorTable.scrollHeight) {
-        donatorWrapper.scrollTop -= 0.5 * donatorTable.scrollHeight;
-      } else {
-        // console.log(donatorWrapper.scrollTop, donatorTable.scrollHeight);
-        donatorWrapper.scrollTop++;
-      }
-    };
-
-    const requesterScroll = () => {
-      // When requester messages are scrolled to the end, the cycle will restart
-      if (requesterWrapper.scrollTop >= 0.5 * requesterWrapper.scrollHeight) {
-        requesterWrapper.scrollTop -= 0.5 * requesterWrapper.scrollHeight;
-      } else {
-        // console.log(requesterWrapper.scrollTop, requesterWrapper.scrollHeight);
-        requesterWrapper.scrollTop++;
-      }
-    };
-    // Set a timer for the scrolling effect
-    let d = setInterval(donatorScroll, 20);
-    let r = setInterval(requesterScroll, 20);
-
-    //When mouseover event is triggered, stop table from scrolling
-    donatorWrapper.addEventListener("mouseover", () => clearInterval(d), false);
-    requesterWrapper.addEventListener(
-      "mouseover",
-      () => clearInterval(r),
-      false
-    );
-    //When mouseout event is triggered, continue scrolling
-    donatorWrapper.addEventListener(
-      "mouseout",
-      () => {
-        d = setInterval(donatorScroll, 20);
-      },
-      false
-    );
-    requesterWrapper.addEventListener(
-      "mouseout",
-      () => {
-        r = setInterval(requesterScroll, 20);
-      },
-      false
-    );
-  }, []);
-
-  // const supportInfoList = [];
-  const list = [];
-  donatorData.forEach((data) => {
-    list.push({
-      name: data["field_1"],
-      content: `捐赠了 ${data["field_12"]} 人民币`,
-      createdAt: data["created_at"],
-    });
-  });
-
-  domesticData.forEach((data) => {
-    list.push({
-      name: data["field_1"],
-      content: `捐赠了 ${data["field_5"][0]} ${data["field_10"]} 个`,
-      createdAt: data["created_at"],
-    });
-  });
-
-  list.sort(function (a, b) {
-    a = new Date(a.createdAt);
-    b = new Date(b.createdAt);
-    return a > b ? -1 : a < b ? 1 : 0;
-  });
-
 
   const hightlightList = [];
   let imagesTemp = [];
@@ -315,11 +250,9 @@ const App = () => {
 
   return (
     <div className="app">
-      
 
       <Header money={money} volunteerCount={volunteerCount} />
 
-      
 
       <section className="support">
         <h1>援助地图</h1>
@@ -338,118 +271,10 @@ const App = () => {
           </div>
         </div>
         <div className="info">
-          <div className="row-1">
-            <a
-              className="btn-mb"
-              rel="noopener noreferrer"
-              href="#helpus"
-              type="button"
-            >
-              提供援助
-            </a>
-            <div id="donator-wrapper" className="w60">
-              <table className="info-table">
-                <tbody>
-                  {list.map((data, index) => (
-                    <tr key={`donator-${index}`}>
-                      <td>{data.name}</td>
-                      <td>{data.content}</td>
-                      <td>{moment(data.createdAt).format("YYYY年MM月DD日")}</td>
-                      {/* <td>{data.createdAt}</td> */}
-                    </tr>
-                  ))}
-                  {/* Repeat the list to make sure the infinite scrolling effect */}
-                  {list.map((data, index) => (
-                    <tr key={`donator-repeat-${index}`}>
-                      <td>{data.name}</td>
-                      <td>{data.content}</td>
-                      <td>{moment(data.createdAt).format("YYYY年MM月DD日")}</td>
-                      {/* <td>{data.createdAt}</td> */}
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
 
-            <div className="w40 dt support-column">
-              <p>我们目前急需口罩和采购口罩的资金。多少不限，请勿坐视。</p>
-              <a
-                className="btn"
-                rel="noopener noreferrer"
-                href="#helpus"
-                type="button"
-              >
-                提供援助
-              </a>
-            </div>
-          </div>
-          <div className="row-2">
-            <div className="w40">
-              <p>
-                凡华人组织和个人都可以通过这里申请口罩援助，我们会认真对待每一条申请。
-              </p>
-              <a
-                className="btn"
-                target="_blank"
-                rel="noopener noreferrer"
-                // href={JINSHUJU.FOREIGN_FORM_LINK}
-                href={"https://jinshuju.net/f/sIDktA"}
-                type="button"
-              >
-                申请援助
-              </a>
-            </div>
-            <div id="requester-wrapper" className="align-right w60">
-              {filteredData.map((data) => (
-                <div className="requester">
-                  <div className="content">{data["field_7"]}</div>
-                  <div className="requester-info">
-                    <p>{`${data["field_11"]}`}</p>
-                    <p>{`${getCountryName(data["field_13"])}-${getCityName(
-                      data["field_15"]
-                    )}`}</p>
-                  </div>
-                </div>
-              ))}
-              {/* Repeat the message list to make sure the infinite scrolling effect */}
-              {filteredData.map((data) => (
-                <div className="requester">
-                  <div className="content">{data["field_7"]}</div>
-                  <div className="requester-info">
-                    <p>{`${data["field_11"]}`}</p>
-                    <p>{`${getCountryName(data["field_13"])}-${getCityName(
-                      data["field_15"]
-                    )}`}</p>
-                  </div>
-                </div>
-              ))}
-            </div>
-            <div className="mb">
-              <p>
-                凡华人组织和个人都可以通过这里申请口罩援助，我们会认真对待每一条申请。
-              </p>
-              <a
-                className="btn"
-                target="_blank"
-                rel="noopener noreferrer"
-                // href={JINSHUJU.FOREIGN_FORM_LINK}
-                href={"https://jinshuju.net/f/sIDktA"}
-                type="button"
-              >
-                申请援助
-              </a>
-              <a
-                className="btn-mb"
-                target="_blank"
-                rel="noopener noreferrer"
-                // href={JINSHUJU.FOREIGN_FORM_LINK}
-                href={"https://jinshuju.net/f/sIDktA"}
-                type="button"
-              >
-                申请援助
-              </a>
-            </div>
-          </div>
+          <Donators {...configDonatorsComponent} />
+
+          <Requesters {...configRequestersComponent} />
         </div>
       </section>
